+++ example/lib/main.dart
@@ -33,6 +33,7 @@ class MyApp extends StatelessWidget {
           // counter didn't reset back to zero; the application is not restarted.
           platform: TargetPlatform.android,
           primarySwatch: Colors.blue,
+          primaryColor: Colors.greenAccent
         ),
         localizationsDelegates: [
           RefreshLocalizations.delegate,